requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
httpie==0.9.9
idna==2.6
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
urllib3==1.22
